Version from baeldung
Version of tcp and udp from baeldung.
I understand it so that's ok.

I actually spent a couple of hours to solve problem.
Problem was in line
out = new PrintWriter(clientSocket.getOutputStream(), true);
We need true. True means autoFlush. We need autoFlush to get result and not waiting for overflow.